refactor: split SortMenu from Librarybar
